refactor : 로또 결과를 출력할 수 있는 기능들을 LottoResult 클래스로 분리
