+++ src/main/kotlin/lotto/Lotto.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import camp.nextstep.edu.missionutils.Randoms
 import javax.swing.JOptionPane.ERROR_MESSAGE
 
 class Lotto(private val numbers: List<Int>) {
@@ -44,6 +45,18 @@ class Lotto(private val numbers: List<Int>) {
             return false
         }
         return true
+    }
+
+    //4. 1~45까지의 숫자 중 중복되지 않는 6개의 숫자를 뽑는다
+    fun getLottoNums() : List<Int> {
+        val numbers = Randoms.pickUniqueNumbersInRange(1, 45, 6)
+        return numbers
+    }
+
+    //3-2. 로또 발행 번호를 오름차순으로 정렬
+    fun arrangePublishNums(publishNums : List<Int>) : List<Int>{
+        val arrangedNums = publishNums.sorted()
+        return arrangedNums
     }
