+++ src/main/kotlin/lotto/LottoExecptions.kt
@@ -1,0 +1,4 @@
+package lotto
+
+class LottoExecptions {
+}
